split out helpers and its spec into separate files; added a few md doc pages
helpers has been split out into comparators, operator_packages, ranged_indexes;
added unit tests for operator packages; added help files for comparators, operator packages, and ranged indexes;
