+++ docs/index.md
@@ -1,33 +1,24 @@
 [![Build Status](https://travis-ci.org/LokiJS-Forge/LokiDB.svg?branch=master)](https://travis-ci.org/LokiJS-Forge/LokiDB)
 [![Coverage Status](https://coveralls.io/repos/github/LokiJS-Forge/LokiDB/badge.svg?branch=feature%2Fadd_build_system)](https://coveralls.io/github/LokiJS-Forge/LokiDB?branch=feature%2Fadd_build_system)
 
+
 # Welcome to LokiDB
+- Javascript Quickstart
+- Typescript Quickstart
 
+# API Reference
+- [Check it out](api/index.html).
 
-Checkout our [API Reference](api/index.html).
+# Topics
+- Query Examples
+- Persistence
+- CollectionTransforms
+- [Understanding Comparators](./comparators.md)
+- [LokiOperatorPackage](./operator_packages.md)
+- [Loki Ranged Indexes](./ranged_indexes.md)
 
-```javascript
-// Runnable code
-console.log("iii");
-console.log("abc");
-console.log(123);
-console.log({a: 2});
-console.log(undefined)
-console.log(null)
-console.log("??")
-console.log($)
-console.log([]);
-console.log({a: [2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6,2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6]});
-console.error("123")
-console.error("123")
-console.error("123")
-console.error("123")
-console.error("123")
-
-```
+# Getting Started
 
 ```javascript
-// Not runnable code
-test = i + 1;
-nothing.call();
+
 ```
