+++ packages/loki/src/collection.ts
@@ -2,7 +2,8 @@ import { LokiEventEmitter } from "./event_emitter";
 import { UniqueIndex } from "./unique_index";
 import { ResultSet } from "./result_set";
 import { DynamicView } from "./dynamic_view";
-import { IRangedIndex, ComparatorMap, RangedIndexFactoryMap } from "./helper";
+import { IRangedIndex, RangedIndexFactoryMap } from "./ranged_indexes";
+import { ComparatorMap } from "./comparators";
 import { clone, CloneMethod } from "./clone";
 import { Doc, Dict } from "../../common/types";
 import { FullTextSearch } from "../../full-text-search/src/full_text_search";
