+++ packages/loki/src/operator_packages.ts
@@ -1,98 +1,8 @@
-/**
- * This file contains LokiOperatorPackages, RangedIndex and Comparator interfaces, as well as
- * global map object instances for registered LokiOperatorPackages, RangedIndex implementations, and Comparator functions
- */
+import { ILokiComparer } from "./comparators";
 
-import { AvlTreeIndex } from "./avl_index";
-
-/* valid range ops that the index must directly support */
-export type RangedValueOperator = "$gt" | "$gte" | "$lt" | "$lte" | "$eq" | "$neq" | "$between";
-
-/* Loki Comparator interface for dependency injection to ranged indexes */
-export interface ILokiComparer<T> {
-  (a: T, b: T): -1 | 0 | 1;
-}
-
-export interface IRangedIndexRequest<T> {
-  op: RangedValueOperator;
-  val: T;
-  high?: T;
-}
-
-export interface IComparatorMap {
-  [name: string]: ILokiComparer<any>;
-}
-
-/** Map/Register of named ILokiComparer functions returning -1, 0, 1 for lt/eq/gt assertions for two passed parameters */
-export let ComparatorMap: IComparatorMap = {
-  "js": CreateJavascriptComparator<any>(),
-  "abstract-js": CreateAbstractJavascriptComparator<any>(),
-  "abstract-date": CreateAbstractDateJavascriptComparator<any>(),
-  "loki": CreateLokiComparator()
-};
-
-/** Hash Interface for global ranged index factory map*/
-export interface IRangedIndexFactoryMap {
-  [name: string]: (name: string, comparator: ILokiComparer<any>) => IRangedIndex<any>;
-}
-
-/** Map/Register of named factory functions returning IRangedIndex instances */
-export let RangedIndexFactoryMap: IRangedIndexFactoryMap = {
-  "avl": (name: string, comparator: ILokiComparer<any>) => { return new AvlTreeIndex(name, comparator); }
-};
-
-/** Defines interface which all loki ranged indexes need to implement */
-export interface IRangedIndex<T> {
-  insert(id: number, val: T): void;
-  update(id: number, val: T): void;
-  remove(id: number): void;
-  restore(tree: any): void;
-  backup(): IRangedIndex<T>;
-
-  rangeRequest(range?: IRangedIndexRequest<T>): number[];
-
-  validateIndex(): boolean;
-}
-
-/** Typescript-friendly factory for strongly typed 'js' comparators */
-export function CreateJavascriptComparator<T>(): ILokiComparer<T> {
-  return (val: T, val2: T) => {
-    if (val === val2) return 0;
-    if (val < val2) return -1;
-    return 1;
-  };
-}
-
-/** Typescript-friendly factory for strongly typed 'abstract js' comparators */
-export function CreateAbstractJavascriptComparator<T>(): ILokiComparer<T> {
-  return (val: T, val2: T) => {
-    if (val == val2) return 0;
-    if (val < val2) return -1;
-    return 1;
-  };
-}
-
-/**
- * Comparator which attempts to deal with deal with dates at comparator level.
- * Should work for dates in any of the object, string, and number formats
- */
-export function CreateAbstractDateJavascriptComparator<T>(): ILokiComparer<T> {
-  return (val: T, val2: T) => {
-    let v1: string = (new Date(val as any).toISOString());
-    let v2: string = (new Date(val2 as any).toISOString());
-    if (v1 == v2) return 0;
-    if (v1 < v2) return -1;
-    return 1;
-  };
-}
-
-/** Typescript-friendly factory for strongly typed 'loki' comparators */
-export function CreateLokiComparator(): ILokiComparer<any> {
-  return (val: any, val2: any) => {
-    if (aeqHelper(val, val2)) return 0;
-    if (ltHelper(val, val2, false)) return -1;
-    return 1;
-  };
+/** Hash interface for named LokiOperatorPackage registration */
+export interface ILokiOperatorPackageMap {
+  [name: string]: LokiOperatorPackage;
 }
 
 /**
@@ -405,11 +315,6 @@ export function sortHelper(prop1: any, prop2: any, descending: boolean): number 
   return 0;
 }
 
-/** Hash interface for named LokiOperatorPackage registration */
-export interface ILokiOperatorPackageMap {
-  [name: string]: LokiOperatorPackage;
-}
-
 /**
  * Default implementation of LokiOperatorPackage, using fastest javascript comparison operators.
  */
@@ -658,6 +563,5 @@ export class ComparatorOperatorPackage<T> extends LokiOperatorPackage {
  */
 export let LokiOperatorPackageMap : ILokiOperatorPackageMap = {
   "js" : new LokiOperatorPackage(),
-  "loki" : new LokiAbstractOperatorPackage(),
-  "comparator" : new ComparatorOperatorPackage<any>(ComparatorMap["loki"])
+  "loki" : new LokiAbstractOperatorPackage()
 };
